modelのconcernsを修正する
