+++ app/models/events/channel_unarchived.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class Events::ChannelUnarchived
-  include CurlBuilder
-  include ChannelBuilder
+  include Builders::CurlBuilder
+  include Builders::ChannelBuilder
 
   def execute(bot_token, params)
     channel = params[:event][:channel]
