+++ app/workers/schedule_message.rb
@@ -4,17 +4,17 @@ class ScheduleMessage
   include Sidekiq::Worker
   sidekiq_options queue: :schedule_message, retry: 5
 
-  include MessageBuilder
+  include Builders::MessageBuilder
   MAX_WAIT_TIME = 5400
 
   def perform(bot_token, member, push_timestamp, message_id)
     message = Message.find(message_id)
 
     if message.image_url
-      scheduled_message = curl_exec(base_url: SlackApiBaseurl::CHAT_SCHEDULE_MESSAGE,
+      scheduled_message = curl_exec(base_url: SlackApis::Baseurl::CHAT_SCHEDULE_MESSAGE,
                                     params: { "token": bot_token, "channel": member, "post_at": push_timestamp, "blocks": schedule_blocks_text_with_image(message) })
     else
-      scheduled_message = curl_exec(base_url: SlackApiBaseurl::CHAT_SCHEDULE_MESSAGE,
+      scheduled_message = curl_exec(base_url: SlackApis::Baseurl::CHAT_SCHEDULE_MESSAGE,
                                     params: { "token": bot_token, "channel": member, "post_at": push_timestamp, "blocks": schedule_blocks_text(message) })
     end
